\name{rmall}
\alias{rmall}
\title{Removes all objects from specified environment}
\usage{
  rmall(env = .GlobalEnv)
}
\arguments{
  \item{env}{Environment from which objects will be
  removed. Default is global environment
  \code{.GlobalEnv}.}
}
\description{
  Removes all objects from specified environment. Default
  is global environment \code{.GlobalEnv}.
}

